# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: VIGIL + ANCHOR service contract
  version: 2.0.0
  description: >-
    INCOMPLETE — x-vuka-implemented reports route registration only, not that
    all specified behavior is complete; only operations marked
    x-vuka-implemented: true are served by the ANCHOR server at this head;
    every other operation is a contract for future work.
    Contract v2 for VIGIL + ANCHOR. Device, guardian and server events are
    signed as specified in docs/VUKA-2-SPEC.md; the contract describes the
    interface and does not itself implement authentication or verification.
servers:
  - url: https://api.example.invalid
    description: PROPOSED placeholder; deployment host is not selected
...
paths:
...
  /v1/events:
    post:
      tags: [ANCHOR]
      operationId: appendEvidenceEvent
      x-vuka-implemented: true
      summary: Append a signed v2 evidence entry
      description: >-
        Device/guardian-signed event submission. The signed details.sig binds the event context;
        X-Vuka-Signature separately authenticates the HTTP request. The server computes chain
        fields, recomputes the payload commitment and stores payload and salt encrypted at rest.
        Check order is request authentication, event_id idempotency lookup, then nonce, counter
        and clock-skew checks. An identical retry returns the original receipt before replay checks.
      security: [{ VukaKeyId: [], VukaTimestamp: [], VukaNonce: [], VukaSignature: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventSubmission' }
      responses:
        '201':
          description: Entry received and appended
          content:
            application/json:
              schema:
                type: object
                required: [event_hash, chain_index, received_at]
                additionalProperties: false
                properties:
                  event_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
                  chain_index: { type: integer, minimum: 0 }
                  received_at: { type: string, format: date-time }
        '400':
          description: Invalid submission shape or payload commitment mismatch.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Genesis registration blocked while simulation-only mode is enabled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                code: simulation_only
                message: genesis registration target must use the sim_ prefix while simulation-only mode is enabled
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
...
  /v1/subjects/{id}/export:
    get:
      tags: [ANCHOR]
      operationId: exportSubjectRecord
      x-vuka-implemented: true
      summary: Export the authenticated subject's permitted evidence prefix
      description: >-
        Subject-authenticated with the §7 signed-request headers and a fresh
        pin_authorised record for action export and this subject id. Missing or
        mismatched PIN authority returns pin_authorisation_required. ADR-0041's
        pre-incident hold returns only genesis through the pre-incident head while
        an incident is open and for 6 hours after its last PIN entry, whichever ends
        later; normal and duress authorisations yield byte-identical visible results.
        Otherwise the full chain is returned. The response is unpaginated. For this
        bodyless request, body_sha256 is SHA-256 of the empty byte string.
        INCOMPLETE — see ADR-0041 T30, implementation pending.
      security: [{ VukaKeyId: [], VukaTimestamp: [], VukaNonce: [], VukaSignature: [] }]
      parameters: [{ $ref: '#/components/parameters/SubjectId' }]
      responses:
        '200':
          description: Subject export with payloads, salts, proofs and receipts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubjectExport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/InsufficientApproval' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    VukaKeyId:
      type: apiKey
      in: header
      name: X-Vuka-Key-Id
      description: Active device or guardian signer_key_id used to resolve the registered public key.
    VukaTimestamp:
      type: apiKey
      in: header
      name: X-Vuka-Ts
      description: RFC 3339 request timestamp included in the signed request statement.
    VukaNonce:
      type: apiKey
      in: header
      name: X-Vuka-Nonce
      description: Per-request nonce; retained per key for 24 hours, subject to event retry ordering in §7.
    VukaSignature:
      type: apiKey
      in: header
      name: X-Vuka-Signature
      description: >-
        Base64 raw P-256 r||s signature over canonical({method, path, ts, body_sha256, nonce});
        body_sha256 hashes the exact HTTP body bytes. For initial registration, verify with the
        submitted signer_pubkey before enrolling the signer_key_id.
    VukaServerSignature:
      type: apiKey
      in: header
      name: X-Vuka-Server-Signature
      description: >-
        Base64 Ed25519 signature over canonical({method, path, ts, body_sha256, nonce});
        body_sha256 hashes the exact HTTP body bytes. Verify only with server_ed25519_public_key
        from contracts/keys/manifest.json; the request cannot supply its own verification key.
  parameters:
...
    SubjectId: { name: id, in: path, required: true, schema: { type: string, minLength: 1 } }
...
  schemas:
...
    EventSubmission:
      type: object
      description: >-
        Device/guardian-submitted v2 event. The server supplies prev_hash, event_hash,
        details.received_at and details.chain_index. It recomputes commitment as
        SHA-256(salt bytes concatenated with canonical(payload)); a mismatch returns 400.
        Payload and salt are sent over TLS and stored separately encrypted at rest.
      required: [action, actor_id, target_type, target_id, ts, details, payload, salt]
      additionalProperties: false
      properties:
        action: { type: string, minLength: 1 }
        actor_id: { type: string, minLength: 1 }
        target_type: { type: string, enum: [journey, subject] }
        target_id: { type: string, minLength: 1 }
        ts: { type: string, format: date-time, description: Signer's source time. }
        details:
          type: object
          required: [v, signer, signer_key_id, counter, event_id, commitment, sig]
          additionalProperties: false
          properties:
            v: { const: 2 }
            signer: { type: string, enum: [device, guardian] }
            signer_key_id: { type: string, minLength: 1 }
            counter: { type: integer, minimum: 0 }
            event_id: { type: string, format: uuid }
            commitment: { type: string, pattern: '^[0-9a-f]{64}$' }
            sig: { type: string, contentEncoding: base64, minLength: 1, description: Event-context signature over the §4 canonical statement. }
            signer_pubkey: { type: string, contentEncoding: base64, description: Present on registration and guardian-accept entries. }
        payload:
          type: object
          description: >-
            Private committed payload. kind and pv select the per-kind schema at
            contracts/payloads/{kind}.v{pv}.json; validate against that schema before acceptance.
            Every value must also be canonical-safe (no floats, unsafe integers or non-ASCII keys).
          required: [kind, pv]
          propertyNames: { pattern: '^[\x00-\x7F]*$' }
          properties:
            kind: { type: string, minLength: 1 }
            pv: { type: integer, minimum: 1 }
          additionalProperties: { $ref: '#/components/schemas/CanonicalJsonValue' }
        salt:
          type: string
          contentEncoding: base64
          minLength: 24
          maxLength: 24
          pattern: '^[A-Za-z0-9+/]{21}[AEIMQUYcgkosw048]==$'
          description: Base64 encoding of exactly 16 random bytes.
    CanonicalJsonValue:
      description: Recursive JSON value subset accepted by the §5 canonical form.
      oneOf:
        - type: string
        - type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        - type: boolean
        - type: 'null'
        - type: array
          items: { $ref: '#/components/schemas/CanonicalJsonValue' }
        - type: object
          propertyNames: { pattern: '^[\x00-\x7F]*$' }
          additionalProperties: { $ref: '#/components/schemas/CanonicalJsonValue' }
    EvidenceEntryV2:
      type: object
      description: >-
        Format-v2 signed chain entry. Genesis prev_hash is 64 zero hex
        characters; it is never null. Receipt time and chain index are server
        fields inside details and are not part of the device signature.
      required: [action, actor_id, target_type, target_id, details, ts, prev_hash, event_hash]
      additionalProperties: false
      properties:
        action: { type: string, minLength: 1 }
        actor_id: { type: string, minLength: 1 }
        target_type: { type: string, enum: [journey, subject] }
        target_id: { type: string, minLength: 1 }
        details:
          type: object
          required: [v, signer, signer_key_id, counter, event_id, commitment, sig, received_at, chain_index]
          additionalProperties: false
          properties:
            v: { const: 2 }
            signer: { type: string, enum: [device, guardian, server] }
            signer_key_id: { type: string, minLength: 1 }
            counter: { type: integer, minimum: 0 }
            event_id: { type: string, format: uuid }
            commitment: { type: string, pattern: '^[0-9a-f]{64}$' }
            sig: { type: string, contentEncoding: base64, minLength: 1 }
            received_at: { type: string, format: date-time }
            chain_index: { type: integer, minimum: 0 }
            signer_pubkey: { type: string, contentEncoding: base64, description: Present on registration entries. }
        ts: { type: string, format: date-time, description: Signer's source time. }
        prev_hash: { type: string, pattern: '^[0-9a-f]{64}$', description: Genesis is exactly 64 zero hex characters; never null. }
        event_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
...
    Proof:
      type: array
      description: Ordered Merkle audit path from leaf to root.
      items:
        type: object
        required: [side, hash]
        additionalProperties: false
        properties:
          side: { type: string, enum: [L, R] }
          hash: { type: string, pattern: '^[0-9a-f]{64}$' }
    AnchorReceipt:
      type: object
      required: [topic_id, sequence_number, consensus_timestamp, running_hash, topic_epoch]
      additionalProperties: false
      properties:
        topic_id: { type: string, minLength: 1 }
        sequence_number: { type: integer, minimum: 1 }
        consensus_timestamp:
          type: string
          pattern: '^[0-9]+\.[0-9]{1,9}$'
          description: Hedera mirror consensus time as seconds.nanoseconds, not RFC 3339.
        running_hash: { type: string, minLength: 1 }
        topic_epoch: { type: integer, minimum: 1, description: Epoch of the pinned topic/key manifest. }
...
    SubjectExport:
      type: object
      description: >-
        Subject-authenticated, fresh-PIN-gated and unpaginated. Contains genesis
        through the permitted head (pre-incident head during ADR-0041's hold,
        current head otherwise), with payloads and salts kept separately from
        the public chain.
      required: [subject_id, entries, payloads, salts, proofs, receipts]
      additionalProperties: false
      properties:
        subject_id: { type: string, minLength: 1 }
        entries: { type: array, items: { $ref: '#/components/schemas/EvidenceEntryV2' } }
        payloads: { type: array, items: { type: object }, description: Private event payloads keyed by event_id. }
        salts: { type: array, items: { type: string, contentEncoding: base64 }, description: Private salts keyed by event_id. }
        proofs: { type: array, items: { $ref: '#/components/schemas/Proof' } }
        receipts: { type: array, items: { $ref: '#/components/schemas/AnchorReceipt' } }
...
    Error:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
  responses:
...
    Unauthorized:
      description: Authentication required or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InsufficientApproval:
      description: Required authenticated human or second distinct signature is absent.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    IdempotencyConflict:
      description: >-
        Idempotency key was reused with a different request shape (`idempotency_conflict`); on
        /sim_bank/v1/release also a hold already released by a different request
        (`hold_already_released`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }